Move Dialog Base to components folder
