+++ src/components/DialogBase/ariaHidder.ts
